Extract read-only queries into exec::queries
