+++ src/server/api/v1/tasks.rs
@@ -17,7 +17,7 @@ use super::AppState;
 use super::LogSource;
 use super::TaskStatus;
 use super::error::Error;
-use crate::system::v1::exec::svc::run_manager::commands;
+use crate::system::v1::exec::queries;
 
 /// Query parameters for listing tasks.
 #[derive(Debug, Clone, Serialize, Deserialize, IntoParams, ToSchema)]
@@ -109,8 +109,8 @@ pub struct GetTaskResponse {
     pub task: Task,
 }
 
-impl From<commands::GetTaskResponse> for GetTaskResponse {
-    fn from(response: commands::GetTaskResponse) -> Self {
+impl From<queries::GetTaskResponse> for GetTaskResponse {
+    fn from(response: queries::GetTaskResponse) -> Self {
         Self {
             task: response.task.into(),
         }
@@ -186,7 +186,7 @@ pub async fn list_tasks(
     };
     let limit = query.limit.unwrap_or(100);
 
-    let response = crate::system::v1::exec::svc::run_manager::list_tasks(
+    let response = queries::list_tasks(
         &state.db,
         query.run_uuid,
         query.status,
@@ -226,7 +226,7 @@ pub async fn get_task(
     State(state): State<AppState>,
     Path(name): Path<String>,
 ) -> Result<Json<GetTaskResponse>, Error> {
-    let response = crate::system::v1::exec::svc::run_manager::get_task(&state.db, name).await?;
+    let response = queries::get_task(&state.db, name).await?;
     Ok(Json(response.into()))
 }
 
@@ -264,14 +264,8 @@ pub async fn get_task_logs(
     };
     let limit = query.limit.unwrap_or(100);
 
-    let response = crate::system::v1::exec::svc::run_manager::get_task_logs(
-        &state.db,
-        name,
-        query.source,
-        query.limit,
-        Some(offset),
-    )
-    .await?;
+    let response =
+        queries::get_task_logs(&state.db, name, query.source, query.limit, Some(offset)).await?;
 
     let next_offset = offset + limit;
     let next_token = if next_offset < response.total {
